+++ src/main.cpp
@@ -1,4 +1,5 @@
 #include "equipment.h"
+#include "json/json.hpp"
 
 using namespace std;
 
@@ -10,23 +11,9 @@ int main() {
     Enchant::Rune      fire_core("fire element resistance", 5);
     sword.mountRune(rune_of_war).mountRune(rune_of_arcane).mountRune(fire_core);
     cout << sword << endl;
-    // Enchant::Attribute a("attack"), b("magic attack", 5.);
-    // cout << a << " " << b << endl;
 
-    // // Can not add () since compiler will recognize it to a function.
-    // Enchant::AttributeTree c;
-    // c.set(b).set(a);
-    // cout << c << endl;
-
-    // // "new" some class is a pointer, so we need to change the class by "*"
-    // c.set(*(new Enchant::Attribute("attack", 6)));
-    // cout << c << endl;
-
-    // Enchant::Rune d(12345678, "attack", 3);
-    // cout << d.getAttribute() << endl;
-
-    // Enchant::Equipment eq(23412385);
-    // eq.mountRune(d);
+    nlohmann::json j = nlohmann::json::parse("[[1001,1],[1002,3.141]]");
+    cout << j;
     system("pause");
     return 0;
 }